* legacy-readout: (168 commits)
  Removed unfinished files (again)
  Old uncommitted updates to sb suppression
  some minor tweaks. support default beammap in yaml
  Debugged flag map issue in sweep.py
  More debugging
  Bug fixes
  Added back a class attribute (only ever used for plotting) removed some lines in main that were used for testing and updated to used the config file
  Updated aligngrid.cfg, shell script for stitching clickthroughs, debugging
  beammap sweep.py updated __main__
  Updates to freqSweep code
  combined some redundant code in beam map sweep.py
  enhancements of beam map clickthrough. Now you can fast forward or rewind past pixel timestreams that pass some checks. Space bar = fastforward, r = rewind
  fixed an erroneous import statement:  from mkidcore.config import config importoldconfig..... --> from mkidcore.config import importoldconfig....
  fixed an erroneous import statement:  from mkidcore.config import config importoldconfig..... --> from mkidcore.config import importoldconfig....
  fixed an erroneous import statement:  from mkidcore.config import config importoldconfig..... --> from mkidcore.config import importoldconfig....
  fixed a bug- some lines got deleted previously on accident, so I replaced them
  Added new autowidesweep script using all powers
  Adding ML interface to new PS/WS data
  add FreqSweep dataclass
  Fixed some plotting bugs for freqSweep()
  ...

+++ mkidreadout/configuration/powersweep/PowerSweepML/PSFitMLData.py
@@ -122,11 +122,11 @@ class PSFitting():
                 'attens':self.Res1.atten1s}
     
     def loadps(self):
-        hd5file=openFile(self.initialFile,mode='r')
-        group = hd5file.getNode('/','r0')
+        hd5file=open_file(self.initialFile,mode='r')
+        group = hd5file.get_node('/','r0')
         self.freq=np.empty(0,dtype='float32')
         
-        for sweep in group._f_walkNodes('Leaf'):
+        for sweep in group._f_walknodes('Leaf'):
             k=sweep.read()
             self.scale = k['scale'][0]
             #print "Scale factor is ", self.scale
@@ -146,6 +146,7 @@ class PSFitMLData():
         self.freqs, self.iq_vels,self.Is,self.Qs, self.attens, self.resIDs = self.get_PS_data()
         self.opt_attens = None
         self.opt_freqs = None
+        self.scores = None
 
     def loadTrainData(self):                
         '''
@@ -201,25 +202,41 @@ class PSFitMLData():
         self.Qs = self.Qs[self.good_res]
         self.resIDs = self.resIDs[self.good_res]
 
-    def savePSTxtFile(self, flag = '', outputFN=None):
+    def savePSTxtFile(self, flag = '', outputFN=None, saveScores=False):
         '''
         Saves a frequency file after inference.  self.opt_attens and self.opt_freqs
         should be populated by an external ML algorithm.
         '''
         if self.opt_attens is None or self.opt_freqs is None:
             raise ValueError('Classify Resonators First!')
+
+        if saveScores:
+            scoreFlag = '_scored'
+        else:
+            scoreFlag = ''
         
-        if outputFN is None: PSSaveFile = self.baseFile + flag + '.txt'
-        else: PSSaveFile = outputFN.rsplit('.',1)[0]+flag+'.txt'
+        if outputFN is None: PSSaveFile = self.baseFile + flag + scoreFlag + '.txt'
+        else: PSSaveFile = outputFN.rsplit('.',1)[0]+flag+scoreFlag+'.txt'
+        
+        if os.path.isfile(PSSaveFile):
+            raise Exception('PS Save data already exists!')
+
         sf = open(PSSaveFile,'wb')
         print 'saving file', PSSaveFile
         print 'baseFile', self.baseFile
         #sf.write('1\t1\t1\t1 \n')
-        for r in range(len(self.opt_attens)):
-            line = "%4i \t %10.9e \t %4i \n" % (self.resIDs[r], self.opt_freqs[r], 
-                                         self.opt_attens[r])
-            sf.write(line)
-            #print line
+        if saveScores:
+            for r in range(len(self.opt_attens)):
+                line = "%4i \t %10.9e \t %4i \t %0.5f \n" % (self.resIDs[r], self.opt_freqs[r], 
+                                             self.opt_attens[r], self.scores[r])
+                sf.write(line)
+                #print line
+        else:
+            for r in range(len(self.opt_attens)):
+                line = "%4i \t %10.9e \t %4i \n" % (self.resIDs[r], self.opt_freqs[r], 
+                                             self.opt_attens[r])
+                sf.write(line)
+                #print line
         sf.close()        
 
     def get_PS_data(self, searchAllRes=True, res_nums=50):
